Added methods to apply filters to product search throug API REST.
An enumeration responsible of filtering the information created and some
get methods created in ProductService.

The only filter created is to find one representative product of every
name.

Now size is string instead of ProducSize type in product and
product-no-images in frontend and a new enum was created to allocate
existing filters.