Change in user creation and  it's possible to get user's wishlist.
Now wishlist icon is displayed only when user is authenticated, but
authentication is still incomplete. Besides the wishlist in now read as
an observable and refreshSession was created in product-card to refresh
the main page everytime someone add or delete an item from the wishlist
(this functionality is not implemented yet).